+++ packages/base/src/page/System/PushNotification/LarkSetting/conponents/ConfigField.tsx
@@ -1,0 +1,39 @@
+import { useTranslation } from 'react-i18next';
+
+import { FormItemLabel } from '@actiontech/shared/lib/components/FormCom';
+import { BasicInput } from '@actiontech/shared';
+
+const ConfigField = () => {
+  const { t } = useTranslation();
+
+  return (
+    <>
+      <FormItemLabel
+        className="has-required-style"
+        label="App ID"
+        name="appKey"
+        rules={[{ required: true }]}
+      >
+        <BasicInput
+          placeholder={t('common.form.placeholder.input', {
+            name: 'App ID'
+          })}
+        />
+      </FormItemLabel>
+      <FormItemLabel
+        className="has-required-style"
+        label="App Secret"
+        name="appSecret"
+        rules={[{ required: true }]}
+      >
+        <BasicInput.Password
+          placeholder={t('common.form.placeholder.input', {
+            name: 'App Secret'
+          })}
+        />
+      </FormItemLabel>
+    </>
+  );
+};
+
+export default ConfigField;
