+++ packages/base/src/page/System/components/ConfigModifyBtn/index.tsx
@@ -1,9 +1,10 @@
+import { useTranslation } from 'react-i18next';
 import { BasicButton, BasicToolTips } from '@actiontech/shared';
 import { IconModify } from '../../../../icon/system';
-import { useTranslation } from 'react-i18next';
 
 const ConfigModifyBtn: React.FC<{ onClick: () => void }> = ({ onClick }) => {
   const { t } = useTranslation();
+
   return (
     <BasicToolTips title={t('common.modify')} titleWidth={54}>
       <BasicButton
