+++ packages/base/src/page/System/components/ConfigTestBtn/index.tsx
@@ -1,6 +1,6 @@
+import { useTranslation } from 'react-i18next';
 import { MutableRefObject } from 'react';
 import { Popover } from 'antd';
-import { useTranslation } from 'react-i18next';
 import { BasicButton, BasicToolTips } from '@actiontech/shared';
 import { IconTest } from '../../../../icon/system';
