+++ packages/base/src/page/System/index.tsx
@@ -3,9 +3,9 @@ import { useDispatch } from 'react-redux';
 import { useTranslation } from 'react-i18next';
 import { Row } from 'antd';
 import { BasicSegmented, PageHeader } from '@actiontech/shared';
-import { ModalName } from '../../data/ModalName';
 import { SystemStyleWrapper } from './style';
 import { initSystemModalStatus } from '../../store/system';
+import { ModalName } from '../../data/ModalName';
 
 // #if [sqle]
 import PushNotification from './PushNotification';
@@ -31,6 +31,7 @@ enum SystemSegmentedKeyEnum {
 
 const System = () => {
   const { t } = useTranslation();
+
   const dispatch = useDispatch();
 
   const options = useMemo(
